Moved gRPC node package root to repo root, made it depend on grpc.gyp
