+++ templates/grpc.gyp.template
@@ -80,7 +80,13 @@
         '-ldl',
         '-lpthread',
         '-lz'
-      ]
+      ],
+      'direct_dependent_settings': {
+        'include_dirs': [
+          '.',
+          'include'
+        ],
+      }
     },
     'targets': [
       % for lib in libs:
